Refactor OpenSSLCipher
The heavily-nested class structure makes it hard to follow what every
class is doing.  Instead, break them out into their own classes by
algorithm.
