+++ common/src/main/java/org/conscrypt/OpenSSLCipher.java
@@ -16,9 +16,6 @@
 
 package org.conscrypt;
 
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
-import java.nio.ByteBuffer;
 import java.security.AlgorithmParameters;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidKeyException;
@@ -43,7 +40,6 @@ import javax.crypto.SecretKey;
 import javax.crypto.ShortBufferException;
 import javax.crypto.spec.IvParameterSpec;
 import javax.crypto.spec.SecretKeySpec;
-import org.conscrypt.NativeRef.EVP_CIPHER_CTX;
 
 /**
  * An implementation of {@link Cipher} using BoringSSL as the backing library.
@@ -474,1020 +470,4 @@ public abstract class OpenSSLCipher extends CipherSpi {
         return encrypting;
     }
 
-    public static abstract class EVP_CIPHER extends OpenSSLCipher {
-        /**
-         * Native pointer for the OpenSSL EVP_CIPHER context.
-         */
-        private final EVP_CIPHER_CTX cipherCtx = new EVP_CIPHER_CTX(
-                NativeCrypto.EVP_CIPHER_CTX_new());
-
-        /**
-         * Whether the cipher has processed any data yet. EVP_CIPHER doesn't
-         * like calling "doFinal()" in decryption mode without processing any
-         * updates.
-         */
-        boolean calledUpdate;
-
-        /**
-         * The block size of the current mode.
-         */
-        private int modeBlockSize;
-
-        public EVP_CIPHER(Mode mode, Padding padding) {
-            super(mode, padding);
-        }
-
-        @Override
-        void engineInitInternal(byte[] encodedKey, AlgorithmParameterSpec params,
-                SecureRandom random) throws InvalidKeyException,
-                InvalidAlgorithmParameterException {
-            byte[] iv;
-            if (params instanceof IvParameterSpec) {
-                IvParameterSpec ivParams = (IvParameterSpec) params;
-                iv = ivParams.getIV();
-            } else {
-                iv = null;
-            }
-
-            final long cipherType = NativeCrypto.EVP_get_cipherbyname(getCipherName(
-                    encodedKey.length, mode));
-            if (cipherType == 0) {
-                throw new InvalidAlgorithmParameterException("Cannot find name for key length = "
-                        + (encodedKey.length * 8) + " and mode = " + mode);
-            }
-
-            final boolean encrypting = isEncrypting();
-
-            final int expectedIvLength = NativeCrypto.EVP_CIPHER_iv_length(cipherType);
-            if (iv == null && expectedIvLength != 0) {
-                if (!encrypting) {
-                    throw new InvalidAlgorithmParameterException("IV must be specified in " + mode
-                            + " mode");
-                }
-
-                iv = new byte[expectedIvLength];
-                if (random != null) {
-                    random.nextBytes(iv);
-                } else {
-                    NativeCrypto.RAND_bytes(iv);
-                }
-            } else if (expectedIvLength == 0 && iv != null) {
-                throw new InvalidAlgorithmParameterException("IV not used in " + mode + " mode");
-            } else if (iv != null && iv.length != expectedIvLength) {
-                throw new InvalidAlgorithmParameterException("expected IV length of "
-                        + expectedIvLength + " but was " + iv.length);
-            }
-
-            this.iv = iv;
-
-            if (supportsVariableSizeKey()) {
-                NativeCrypto.EVP_CipherInit_ex(cipherCtx, cipherType, null, null, encrypting);
-                NativeCrypto.EVP_CIPHER_CTX_set_key_length(cipherCtx, encodedKey.length);
-                NativeCrypto.EVP_CipherInit_ex(cipherCtx, 0, encodedKey, iv, isEncrypting());
-            } else {
-                NativeCrypto.EVP_CipherInit_ex(cipherCtx, cipherType, encodedKey, iv, encrypting);
-            }
-
-            // OpenSSL only supports PKCS5 Padding.
-            NativeCrypto
-                    .EVP_CIPHER_CTX_set_padding(cipherCtx, getPadding() == Padding.PKCS5PADDING);
-            modeBlockSize = NativeCrypto.EVP_CIPHER_CTX_block_size(cipherCtx);
-            calledUpdate = false;
-        }
-
-        @Override
-        int updateInternal(byte[] input, int inputOffset, int inputLen, byte[] output,
-                int outputOffset, int maximumLen) throws ShortBufferException {
-            final int intialOutputOffset = outputOffset;
-
-            final int bytesLeft = output.length - outputOffset;
-            if (bytesLeft < maximumLen) {
-                throw new ShortBufferException("output buffer too small during update: "
-                        + bytesLeft + " < " + maximumLen);
-            }
-
-            outputOffset += NativeCrypto.EVP_CipherUpdate(cipherCtx, output, outputOffset, input,
-                    inputOffset, inputLen);
-
-            calledUpdate = true;
-
-            return outputOffset - intialOutputOffset;
-        }
-
-        @Override
-        int doFinalInternal(byte[] output, int outputOffset, int maximumLen)
-                throws IllegalBlockSizeException, BadPaddingException, ShortBufferException {
-            /* Remember this so we can tell how many characters were written. */
-            final int initialOutputOffset = outputOffset;
-
-            /*
-             * If we're decrypting and haven't had any input, we should return
-             * null. Otherwise OpenSSL will complain if we call final.
-             */
-            if (!isEncrypting() && !calledUpdate) {
-                return 0;
-            }
-
-            /* Allow OpenSSL to pad if necessary and clean up state. */
-            final int bytesLeft = output.length - outputOffset;
-            final int writtenBytes;
-            if (bytesLeft >= maximumLen) {
-                writtenBytes = NativeCrypto.EVP_CipherFinal_ex(cipherCtx, output, outputOffset);
-            } else {
-                final byte[] lastBlock = new byte[maximumLen];
-                writtenBytes = NativeCrypto.EVP_CipherFinal_ex(cipherCtx, lastBlock, 0);
-                if (writtenBytes > bytesLeft) {
-                    throw new ShortBufferException("buffer is too short: " + writtenBytes + " > "
-                            + bytesLeft);
-                } else if (writtenBytes > 0) {
-                    System.arraycopy(lastBlock, 0, output, outputOffset, writtenBytes);
-                }
-            }
-            outputOffset += writtenBytes;
-
-            reset();
-
-            return outputOffset - initialOutputOffset;
-        }
-
-        @Override
-        int getOutputSizeForFinal(int inputLen) {
-            if (modeBlockSize == 1) {
-                return inputLen;
-            } else {
-                final int buffered = NativeCrypto.get_EVP_CIPHER_CTX_buf_len(cipherCtx);
-
-                if (getPadding() == Padding.NOPADDING) {
-                    return buffered + inputLen;
-                } else {
-                    final boolean finalUsed = NativeCrypto.get_EVP_CIPHER_CTX_final_used(cipherCtx);
-                    // There is an additional buffer containing the possible final block.
-                    int totalLen = inputLen + buffered + (finalUsed ? modeBlockSize : 0);
-                    // Extra block for remainder bytes plus padding.
-                    // In case it's encrypting and there are no remainder bytes, add an extra block
-                    // consisting only of padding.
-                    totalLen += ((totalLen % modeBlockSize != 0) || isEncrypting())
-                            ? modeBlockSize : 0;
-                    // The minimum multiple of {@code modeBlockSize} that can hold all the bytes.
-                    return totalLen - (totalLen % modeBlockSize);
-                }
-            }
-        }
-
-        @Override
-        int getOutputSizeForUpdate(int inputLen) {
-            return getOutputSizeForFinal(inputLen);
-        }
-
-        /**
-         * Returns the OpenSSL cipher name for the particular {@code keySize}
-         * and cipher {@code mode}.
-         */
-        abstract String getCipherName(int keySize, Mode mode);
-
-        /**
-         * Reset this Cipher instance state to process a new chunk of data.
-         */
-        private void reset() {
-            NativeCrypto.EVP_CipherInit_ex(cipherCtx, 0, encodedKey, iv, isEncrypting());
-            calledUpdate = false;
-        }
-
-        abstract static class AES_BASE extends EVP_CIPHER {
-            private static final int AES_BLOCK_SIZE = 16;
-
-            AES_BASE(Mode mode, Padding padding) {
-                super(mode, padding);
-            }
-
-            @Override
-            void checkSupportedMode(Mode mode) throws NoSuchAlgorithmException {
-                switch (mode) {
-                    case CBC:
-                    case CTR:
-                    case ECB:
-                        return;
-                    default:
-                        throw new NoSuchAlgorithmException("Unsupported mode " + mode.toString());
-                }
-            }
-
-            @Override
-            void checkSupportedPadding(Padding padding) throws NoSuchPaddingException {
-                switch (padding) {
-                    case NOPADDING:
-                    case PKCS5PADDING:
-                        return;
-                    default:
-                        throw new NoSuchPaddingException(
-                                "Unsupported padding " + padding.toString());
-                }
-            }
-
-            @Override
-            String getBaseCipherName() {
-                return "AES";
-            }
-
-            @Override
-            String getCipherName(int keyLength, Mode mode) {
-                return "aes-" + (keyLength * 8) + "-" + mode.toString().toLowerCase(Locale.US);
-            }
-
-            @Override
-            int getCipherBlockSize() {
-                return AES_BLOCK_SIZE;
-            }
-        }
-
-        public static class AES extends AES_BASE {
-            AES(Mode mode, Padding padding) {
-                super(mode, padding);
-            }
-
-            public static class CBC extends AES {
-                public CBC(Padding padding) {
-                    super(Mode.CBC, padding);
-                }
-
-                public static class NoPadding extends CBC {
-                    public NoPadding() {
-                        super(Padding.NOPADDING);
-                    }
-                }
-
-                public static class PKCS5Padding extends CBC {
-                    public PKCS5Padding() {
-                        super(Padding.PKCS5PADDING);
-                    }
-                }
-            }
-
-            public static class CTR extends AES {
-                public CTR() {
-                    super(Mode.CTR, Padding.NOPADDING);
-                }
-            }
-
-            public static class ECB extends AES {
-                public ECB(Padding padding) {
-                    super(Mode.ECB, padding);
-                }
-
-                public static class NoPadding extends ECB {
-                    public NoPadding() {
-                        super(Padding.NOPADDING);
-                    }
-                }
-
-                public static class PKCS5Padding extends ECB {
-                    public PKCS5Padding() {
-                        super(Padding.PKCS5PADDING);
-                    }
-                }
-            }
-
-            @Override
-            void checkSupportedKeySize(int keyLength) throws InvalidKeyException {
-                switch (keyLength) {
-                    case 16: // AES 128
-                    case 24: // AES 192
-                    case 32: // AES 256
-                        return;
-                    default:
-                        throw new InvalidKeyException("Unsupported key size: " + keyLength
-                                + " bytes");
-                }
-            }
-        }
-
-        public static class AES_128 extends AES_BASE {
-            AES_128(Mode mode, Padding padding) {
-                super(mode, padding);
-            }
-
-            public static class CBC extends AES_128 {
-                public CBC(Padding padding) {
-                    super(Mode.CBC, padding);
-                }
-
-                public static class NoPadding extends CBC {
-                    public NoPadding() {
-                        super(Padding.NOPADDING);
-                    }
-                }
-
-                public static class PKCS5Padding extends CBC {
-                    public PKCS5Padding() {
-                        super(Padding.PKCS5PADDING);
-                    }
-                }
-            }
-
-            public static class CTR extends AES_128 {
-                public CTR() {
-                    super(Mode.CTR, Padding.NOPADDING);
-                }
-            }
-
-            public static class ECB extends AES_128 {
-                public ECB(Padding padding) {
-                    super(Mode.ECB, padding);
-                }
-
-                public static class NoPadding extends ECB {
-                    public NoPadding() {
-                        super(Padding.NOPADDING);
-                    }
-                }
-
-                public static class PKCS5Padding extends ECB {
-                    public PKCS5Padding() {
-                        super(Padding.PKCS5PADDING);
-                    }
-                }
-            }
-
-            @Override
-            void checkSupportedKeySize(int keyLength) throws InvalidKeyException {
-                if (keyLength != 16) { // 128 bits
-                    throw new InvalidKeyException("Unsupported key size: " + keyLength + " bytes");
-                }
-            }
-        }
-
-        public static class AES_256 extends AES_BASE {
-            AES_256(Mode mode, Padding padding) {
-                super(mode, padding);
-            }
-
-            public static class CBC extends AES_256 {
-                public CBC(Padding padding) {
-                    super(Mode.CBC, padding);
-                }
-
-                public static class NoPadding extends CBC {
-                    public NoPadding() {
-                        super(Padding.NOPADDING);
-                    }
-                }
-
-                public static class PKCS5Padding extends CBC {
-                    public PKCS5Padding() {
-                        super(Padding.PKCS5PADDING);
-                    }
-                }
-            }
-
-            public static class CTR extends AES_256 {
-                public CTR() {
-                    super(Mode.CTR, Padding.NOPADDING);
-                }
-            }
-
-            public static class ECB extends AES_256 {
-                public ECB(Padding padding) {
-                    super(Mode.ECB, padding);
-                }
-
-                public static class NoPadding extends ECB {
-                    public NoPadding() {
-                        super(Padding.NOPADDING);
-                    }
-                }
-
-                public static class PKCS5Padding extends ECB {
-                    public PKCS5Padding() {
-                        super(Padding.PKCS5PADDING);
-                    }
-                }
-            }
-
-            @Override
-            void checkSupportedKeySize(int keyLength) throws InvalidKeyException {
-                if (keyLength != 32) { // 256 bits
-                    throw new InvalidKeyException("Unsupported key size: " + keyLength + " bytes");
-                }
-            }
-        }
-
-        public static class DESEDE extends EVP_CIPHER {
-            private static final int DES_BLOCK_SIZE = 8;
-
-            public DESEDE(Mode mode, Padding padding) {
-                super(mode, padding);
-            }
-
-            public static class CBC extends DESEDE {
-                public CBC(Padding padding) {
-                    super(Mode.CBC, padding);
-                }
-
-                public static class NoPadding extends CBC {
-                    public NoPadding() {
-                        super(Padding.NOPADDING);
-                    }
-                }
-
-                public static class PKCS5Padding extends CBC {
-                    public PKCS5Padding() {
-                        super(Padding.PKCS5PADDING);
-                    }
-                }
-            }
-
-            @Override
-            String getBaseCipherName() {
-                return "DESede";
-            }
-
-            @Override
-            String getCipherName(int keySize, Mode mode) {
-                final String baseCipherName;
-                if (keySize == 16) {
-                    baseCipherName = "des-ede";
-                } else {
-                    baseCipherName = "des-ede3";
-                }
-
-                return baseCipherName + "-" + mode.toString().toLowerCase(Locale.US);
-            }
-
-            @Override
-            void checkSupportedKeySize(int keySize) throws InvalidKeyException {
-                if (keySize != 16 && keySize != 24) {
-                    throw new InvalidKeyException("key size must be 128 or 192 bits");
-                }
-            }
-
-            @Override
-            void checkSupportedMode(Mode mode) throws NoSuchAlgorithmException {
-                if (mode != Mode.CBC) {
-                    throw new NoSuchAlgorithmException("Unsupported mode " + mode.toString());
-                }
-            }
-
-            @Override
-            void checkSupportedPadding(Padding padding) throws NoSuchPaddingException {
-                switch (padding) {
-                    case NOPADDING:
-                    case PKCS5PADDING:
-                        return;
-                    default:
-                        throw new NoSuchPaddingException("Unsupported padding "
-                                + padding.toString());
-                }
-            }
-
-            @Override
-            int getCipherBlockSize() {
-                return DES_BLOCK_SIZE;
-            }
-        }
-
-        public static class ARC4 extends EVP_CIPHER {
-            public ARC4() {
-                // Modes and padding don't make sense for ARC4.
-                super(Mode.ECB, Padding.NOPADDING);
-            }
-
-            @Override
-            String getBaseCipherName() {
-                return "ARCFOUR";
-            }
-
-            @Override
-            String getCipherName(int keySize, Mode mode) {
-                return "rc4";
-            }
-
-            @Override
-            void checkSupportedKeySize(int keySize) throws InvalidKeyException {
-            }
-
-            @Override
-            void checkSupportedMode(Mode mode) throws NoSuchAlgorithmException {
-                if (mode != Mode.NONE && mode != Mode.ECB) {
-                    throw new NoSuchAlgorithmException("Unsupported mode " + mode.toString());
-                }
-            }
-
-            @Override
-            void checkSupportedPadding(Padding padding) throws NoSuchPaddingException {
-                if (padding != Padding.NOPADDING) {
-                    throw new NoSuchPaddingException("Unsupported padding " + padding.toString());
-                }
-            }
-
-            @Override
-            int getCipherBlockSize() {
-                return 0;
-            }
-
-            @Override
-            boolean supportsVariableSizeKey() {
-                return true;
-            }
-        }
-    }
-
-    public static abstract class EVP_AEAD extends OpenSSLCipher {
-        /**
-         * The default tag size when one is not specified. Default to
-         * full-length tags (128-bits or 16 octets).
-         */
-        private static final int DEFAULT_TAG_SIZE_BITS = 16 * 8;
-
-        /**
-         * Keeps track of the last used block size.
-         */
-        private static int lastGlobalMessageSize = 32;
-
-        /**
-         * The previously used key to prevent key + nonce (IV) reuse.
-         */
-        private byte[] previousKey;
-
-        /**
-         * The previously used nonce (IV) to prevent key + nonce reuse.
-         */
-        private byte[] previousIv;
-
-        /**
-         * When set this instance must be initialized before use again. This prevents key
-         * and IV reuse.
-         */
-        private boolean mustInitialize;
-
-        /**
-         * The byte array containing the bytes written.
-         */
-        byte[] buf;
-
-        /**
-         * The number of bytes written.
-         */
-        int bufCount;
-
-        /**
-         * AEAD cipher reference.
-         */
-        long evpAead;
-
-        /**
-         * Additional authenticated data.
-         */
-        private byte[] aad;
-
-        /**
-         * The length of the AEAD cipher tag in bytes.
-         */
-        int tagLengthInBytes;
-
-        public EVP_AEAD(Mode mode) {
-            super(mode, Padding.NOPADDING);
-        }
-
-        private void checkInitialization() {
-            if (mustInitialize) {
-                throw new IllegalStateException(
-                        "Cannot re-use same key and IV for multiple encryptions");
-            }
-        }
-
-        /** Constant-time array comparison.  Since we are using this to compare keys, we want to
-         * ensure there's no opportunity for a timing attack. */
-        private boolean arraysAreEqual(byte[] a, byte[] b) {
-            if (a.length != b.length) {
-                return false;
-            }
-
-            int diff = 0;
-            for (int i = 0; i < a.length; i++) {
-                diff |= a[i] ^ b[i];
-            }
-            return diff == 0;
-        }
-
-        private void expand(int i) {
-            /* Can the buffer handle i more bytes, if not expand it */
-            if (bufCount + i <= buf.length) {
-                return;
-            }
-
-            byte[] newbuf = new byte[(bufCount + i) * 2];
-            System.arraycopy(buf, 0, newbuf, 0, bufCount);
-            buf = newbuf;
-        }
-
-        private void reset() {
-            aad = null;
-            final int lastBufSize = lastGlobalMessageSize;
-            if (buf == null) {
-                buf = new byte[lastBufSize];
-            } else if (bufCount > 0 && bufCount != lastBufSize) {
-                lastGlobalMessageSize = bufCount;
-                if (buf.length != bufCount) {
-                    buf = new byte[bufCount];
-                }
-            }
-            bufCount = 0;
-        }
-
-        @Override
-        void engineInitInternal(byte[] encodedKey, AlgorithmParameterSpec params,
-                SecureRandom random) throws InvalidKeyException,
-                InvalidAlgorithmParameterException {
-            byte[] iv;
-            final int tagLenBits;
-            if (params == null) {
-                iv = null;
-                tagLenBits = DEFAULT_TAG_SIZE_BITS;
-            } else {
-                GCMParameters gcmParams = Platform.fromGCMParameterSpec(params);
-                if (gcmParams != null) {
-                    iv = gcmParams.getIV();
-                    tagLenBits = gcmParams.getTLen();
-                } else if (params instanceof IvParameterSpec) {
-                    IvParameterSpec ivParams = (IvParameterSpec) params;
-                    iv = ivParams.getIV();
-                    tagLenBits = DEFAULT_TAG_SIZE_BITS;
-                } else {
-                    iv = null;
-                    tagLenBits = DEFAULT_TAG_SIZE_BITS;
-                }
-            }
-
-            if (tagLenBits % 8 != 0) {
-                throw new InvalidAlgorithmParameterException(
-                        "Tag length must be a multiple of 8; was " + tagLengthInBytes);
-            }
-
-            tagLengthInBytes = tagLenBits / 8;
-
-            final boolean encrypting = isEncrypting();
-
-            evpAead = getEVP_AEAD(encodedKey.length);
-
-            final int expectedIvLength = NativeCrypto.EVP_AEAD_nonce_length(evpAead);
-            if (iv == null && expectedIvLength != 0) {
-                if (!encrypting) {
-                    throw new InvalidAlgorithmParameterException("IV must be specified in " + mode
-                            + " mode");
-                }
-
-                iv = new byte[expectedIvLength];
-                if (random != null) {
-                    random.nextBytes(iv);
-                } else {
-                    NativeCrypto.RAND_bytes(iv);
-                }
-            } else if (expectedIvLength == 0 && iv != null) {
-                throw new InvalidAlgorithmParameterException("IV not used in " + mode + " mode");
-            } else if (iv != null && iv.length != expectedIvLength) {
-                throw new InvalidAlgorithmParameterException("Expected IV length of "
-                        + expectedIvLength + " but was " + iv.length);
-            }
-
-            if (isEncrypting() && iv != null) {
-                if (previousKey != null && previousIv != null
-                        && arraysAreEqual(previousKey, encodedKey)
-                        && arraysAreEqual(previousIv, iv)) {
-                    mustInitialize = true;
-                    throw new InvalidAlgorithmParameterException(
-                            "When using AEAD key and IV must not be re-used");
-                }
-
-                this.previousKey = encodedKey;
-                this.previousIv = iv;
-            }
-            mustInitialize = false;
-            this.iv = iv;
-            reset();
-        }
-
-        @Override
-        protected int engineDoFinal(byte[] input, int inputOffset, int inputLen, byte[] output,
-                int outputOffset) throws ShortBufferException, IllegalBlockSizeException,
-                BadPaddingException {
-            // Because the EVP_AEAD updateInternal processes input but doesn't create any output
-            // (and thus can't check the output buffer), we need to add this check before the
-            // superclass' processing to ensure that updateInternal is never called if the
-            // output buffer isn't large enough.
-            if (output != null) {
-                if (getOutputSizeForFinal(inputLen) > output.length - outputOffset) {
-                    throw new ShortBufferException("Insufficient output space");
-                }
-            }
-            return super.engineDoFinal(input, inputOffset, inputLen, output, outputOffset);
-        }
-
-        @Override
-        int updateInternal(byte[] input, int inputOffset, int inputLen, byte[] output,
-                int outputOffset, int maximumLen) throws ShortBufferException {
-            checkInitialization();
-            if (buf == null) {
-                throw new IllegalStateException("Cipher not initialized");
-            }
-
-            ArrayUtils.checkOffsetAndCount(input.length, inputOffset, inputLen);
-            if (inputLen > 0) {
-                expand(inputLen);
-                System.arraycopy(input, inputOffset, buf, this.bufCount, inputLen);
-                this.bufCount += inputLen;
-            }
-            return 0;
-        }
-
-        @SuppressWarnings("LiteralClassName")
-        private void throwAEADBadTagExceptionIfAvailable(String message, Throwable cause)
-                throws BadPaddingException {
-            Constructor<?> aeadBadTagConstructor;
-            try {
-                aeadBadTagConstructor = Class.forName("javax.crypto.AEADBadTagException")
-                                                .getConstructor(String.class);
-            } catch (Exception ignored) {
-                return;
-            }
-
-            BadPaddingException badTagException = null;
-            try {
-                badTagException = (BadPaddingException) aeadBadTagConstructor.newInstance(message);
-                badTagException.initCause(cause);
-            } catch (IllegalAccessException e2) {
-                // Fall through
-            } catch (InstantiationException e2) {
-                // Fall through
-            } catch (InvocationTargetException e2) {
-                throw(BadPaddingException) new BadPaddingException().initCause(
-                        e2.getTargetException());
-            }
-            if (badTagException != null) {
-                throw badTagException;
-            }
-        }
-
-        @Override
-        int doFinalInternal(byte[] output, int outputOffset, int maximumLen)
-                throws ShortBufferException, IllegalBlockSizeException, BadPaddingException {
-            checkInitialization();
-            final int bytesWritten;
-            try {
-                if (isEncrypting()) {
-                    bytesWritten = NativeCrypto.EVP_AEAD_CTX_seal(evpAead, encodedKey,
-                            tagLengthInBytes, output, outputOffset, iv, buf, 0, bufCount, aad);
-                } else {
-                    bytesWritten = NativeCrypto.EVP_AEAD_CTX_open(evpAead, encodedKey,
-                            tagLengthInBytes, output, outputOffset, iv, buf, 0, bufCount, aad);
-                }
-            } catch (BadPaddingException e) {
-                throwAEADBadTagExceptionIfAvailable(e.getMessage(), e.getCause());
-                throw e;
-            }
-            if (isEncrypting()) {
-                mustInitialize = true;
-            }
-            reset();
-            return bytesWritten;
-        }
-
-        @Override
-        void checkSupportedPadding(Padding padding) throws NoSuchPaddingException {
-            if (padding != Padding.NOPADDING) {
-                throw new NoSuchPaddingException("Must be NoPadding for AEAD ciphers");
-            }
-        }
-
-        /**
-         * AEAD buffers everything until a final output.
-         */
-        @Override
-        int getOutputSizeForUpdate(int inputLen) {
-            return 0;
-        }
-
-        @Override
-        int getOutputSizeForFinal(int inputLen) {
-            return bufCount + inputLen
-                    + (isEncrypting() ? NativeCrypto.EVP_AEAD_max_overhead(evpAead) : 0);
-        }
-
-        // Intentionally missing Override to compile on old versions of Android
-        @SuppressWarnings("MissingOverride")
-        protected void engineUpdateAAD(byte[] input, int inputOffset, int inputLen) {
-            checkInitialization();
-            if (aad == null) {
-                aad = Arrays.copyOfRange(input, inputOffset, inputOffset + inputLen);
-            } else {
-                int newSize = aad.length + inputLen;
-                byte[] newaad = new byte[newSize];
-                System.arraycopy(aad, 0, newaad, 0, aad.length);
-                System.arraycopy(input, inputOffset, newaad, aad.length, inputLen);
-                aad = newaad;
-            }
-        }
-
-        // Intentionally missing Override to compile on old versions of Android
-        @SuppressWarnings("MissingOverride")
-        protected void engineUpdateAAD(ByteBuffer buf) {
-            checkInitialization();
-            if (aad == null) {
-                aad = new byte[buf.remaining()];
-                buf.get(aad);
-            } else {
-                int newSize = aad.length + buf.remaining();
-                byte[] newaad = new byte[newSize];
-                System.arraycopy(aad, 0, newaad, 0, aad.length);
-                buf.get(newaad, aad.length, buf.remaining());
-                aad = newaad;
-            }
-        }
-
-        abstract long getEVP_AEAD(int keyLength) throws InvalidKeyException;
-
-        public abstract static class AES extends EVP_AEAD {
-            private static final int AES_BLOCK_SIZE = 16;
-
-            AES(Mode mode) {
-                super(mode);
-            }
-
-            @Override
-            void checkSupportedKeySize(int keyLength) throws InvalidKeyException {
-                switch (keyLength) {
-                    case 16: // AES 128
-                    case 32: // AES 256
-                        return;
-                    default:
-                        throw new InvalidKeyException("Unsupported key size: " + keyLength
-                                + " bytes (must be 16 or 32)");
-                }
-            }
-
-            @Override
-            String getBaseCipherName() {
-                return "AES";
-            }
-
-            @Override
-            int getCipherBlockSize() {
-                return AES_BLOCK_SIZE;
-            }
-
-            public static class GCM extends AES {
-
-                public GCM() {
-                    super(Mode.GCM);
-                }
-
-                @Override
-                void checkSupportedMode(Mode mode) throws NoSuchAlgorithmException {
-                    if (mode != Mode.GCM) {
-                        throw new NoSuchAlgorithmException("Mode must be GCM");
-                    }
-                }
-
-                @Override
-                protected AlgorithmParameters engineGetParameters() {
-                    // iv will be non-null after initialization.
-                    if (iv == null) {
-                        return null;
-                    }
-
-                    AlgorithmParameterSpec spec = Platform.toGCMParameterSpec(
-                            tagLengthInBytes * 8, iv);
-                    if (spec == null) {
-                        // The platform doesn't support GCMParameterSpec. Fall back to
-                        // the generic AES parameters so at least the caller can get the
-                        // IV.
-                        return super.engineGetParameters();
-                    }
-
-                    try {
-                        AlgorithmParameters params = AlgorithmParameters.getInstance("GCM");
-                        params.init(spec);
-                        return params;
-                    } catch (NoSuchAlgorithmException e) {
-                        // We should not get here.
-                        throw (Error) new AssertionError("GCM not supported").initCause(e);
-                    } catch (InvalidParameterSpecException e) {
-                        // This may happen since Conscrypt doesn't provide this itself.
-                        return null;
-                    }
-                }
-
-                @Override
-                protected AlgorithmParameterSpec getParameterSpec(AlgorithmParameters params)
-                        throws InvalidAlgorithmParameterException {
-                    if (params != null) {
-                        AlgorithmParameterSpec spec = Platform.fromGCMParameters(params);
-                        if (spec != null) {
-                            return spec;
-                        }
-                        return super.getParameterSpec(params);
-                    }
-                    return null;
-                }
-
-                @Override
-                long getEVP_AEAD(int keyLength) throws InvalidKeyException {
-                    if (keyLength == 16) {
-                        return NativeCrypto.EVP_aead_aes_128_gcm();
-                    } else if (keyLength == 32) {
-                        return NativeCrypto.EVP_aead_aes_256_gcm();
-                    } else {
-                        throw new RuntimeException("Unexpected key length: " + keyLength);
-                    }
-                }
-
-                @Override
-                int getOutputSizeForFinal(int inputLen) {
-                    // For GCM, the tag is a fixed length and there is no padding or other
-                    // concerns, so we can calculate the exact length required without a
-                    // native call
-                    if (isEncrypting()) {
-                        return bufCount + inputLen + tagLengthInBytes;
-                    } else {
-                        return Math.max(0, bufCount + inputLen - tagLengthInBytes);
-                    }
-                }
-
-                public static class AES_128 extends GCM {
-                    public AES_128() {}
-
-                    @Override
-                    void checkSupportedKeySize(int keyLength) throws InvalidKeyException {
-                        if (keyLength != 16) { // 128 bits
-                            throw new InvalidKeyException(
-                                    "Unsupported key size: " + keyLength + " bytes (must be 16)");
-                        }
-                    }
-                }
-
-                public static class AES_256 extends GCM {
-                    public AES_256() {}
-
-                    @Override
-                    void checkSupportedKeySize(int keyLength) throws InvalidKeyException {
-                        if (keyLength != 32) { // 256 bits
-                            throw new InvalidKeyException(
-                                    "Unsupported key size: " + keyLength + " bytes (must be 32)");
-                        }
-                    }
-                }
-            }
-        }
-
-        public static class ChaCha20 extends EVP_AEAD {
-            public ChaCha20() {
-                super(Mode.POLY1305);
-            }
-
-            @Override
-            void checkSupportedKeySize(int keyLength) throws InvalidKeyException {
-                if (keyLength != 32) {
-                    throw new InvalidKeyException("Unsupported key size: " + keyLength
-                            + " bytes (must be 32)");
-                }
-            }
-
-            @Override
-            String getBaseCipherName() {
-                return "ChaCha20";
-            }
-
-            @Override
-            int getCipherBlockSize() {
-                return 0;
-            }
-
-            @Override
-            void checkSupportedMode(Mode mode) throws NoSuchAlgorithmException {
-                if (mode != Mode.POLY1305) {
-                    throw new NoSuchAlgorithmException("Mode must be Poly1305");
-                }
-            }
-
-            @Override
-            long getEVP_AEAD(int keyLength) throws InvalidKeyException {
-                if (keyLength == 32) {
-                    return NativeCrypto.EVP_aead_chacha20_poly1305();
-                } else {
-                    throw new RuntimeException("Unexpected key length: " + keyLength);
-                }
-            }
-
-            @Override
-            int getOutputSizeForFinal(int inputLen) {
-                // For ChaCha20+Poly1305, the tag is always 16 bytes long and there is no
-                // padding or other concerns, so we can calculate the exact length required
-                // without a native call
-                if (isEncrypting()) {
-                    return bufCount + inputLen + 16;
-                } else {
-                    return Math.max(0, bufCount + inputLen - 16);
-                }
-            }
-        }
-    }
 }
