+++ common/src/main/java/org/conscrypt/OpenSSLProvider.java
@@ -356,26 +356,26 @@ public final class OpenSSLProvider extends Provider {
          * alias for PKCS5Padding.
          */
         putSymmetricCipherImplClass("AES/ECB/NoPadding",
-                "OpenSSLCipher$EVP_CIPHER$AES$ECB$NoPadding");
+                "OpenSSLEvpCipherAES$AES$ECB$NoPadding");
         putSymmetricCipherImplClass("AES/ECB/PKCS5Padding",
-                "OpenSSLCipher$EVP_CIPHER$AES$ECB$PKCS5Padding");
+                "OpenSSLEvpCipherAES$AES$ECB$PKCS5Padding");
         put("Alg.Alias.Cipher.AES/ECB/PKCS7Padding", "AES/ECB/PKCS5Padding");
         putSymmetricCipherImplClass("AES/CBC/NoPadding",
-                "OpenSSLCipher$EVP_CIPHER$AES$CBC$NoPadding");
+                "OpenSSLEvpCipherAES$AES$CBC$NoPadding");
         putSymmetricCipherImplClass("AES/CBC/PKCS5Padding",
-                "OpenSSLCipher$EVP_CIPHER$AES$CBC$PKCS5Padding");
+                "OpenSSLEvpCipherAES$AES$CBC$PKCS5Padding");
         put("Alg.Alias.Cipher.AES/CBC/PKCS7Padding", "AES/CBC/PKCS5Padding");
-        putSymmetricCipherImplClass("AES/CTR/NoPadding", "OpenSSLCipher$EVP_CIPHER$AES$CTR");
+        putSymmetricCipherImplClass("AES/CTR/NoPadding", "OpenSSLEvpCipherAES$AES$CTR");
 
         putSymmetricCipherImplClass(
-                "AES_128/ECB/NoPadding", "OpenSSLCipher$EVP_CIPHER$AES_128$ECB$NoPadding");
+                "AES_128/ECB/NoPadding", "OpenSSLEvpCipherAES$AES_128$ECB$NoPadding");
         putSymmetricCipherImplClass(
-                "AES_128/ECB/PKCS5Padding", "OpenSSLCipher$EVP_CIPHER$AES_128$ECB$PKCS5Padding");
+                "AES_128/ECB/PKCS5Padding", "OpenSSLEvpCipherAES$AES_128$ECB$PKCS5Padding");
         put("Alg.Alias.Cipher.AES_128/ECB/PKCS7Padding", "AES_128/ECB/PKCS5Padding");
         putSymmetricCipherImplClass(
-                "AES_128/CBC/NoPadding", "OpenSSLCipher$EVP_CIPHER$AES_128$CBC$NoPadding");
+                "AES_128/CBC/NoPadding", "OpenSSLEvpCipherAES$AES_128$CBC$NoPadding");
         putSymmetricCipherImplClass(
-                "AES_128/CBC/PKCS5Padding", "OpenSSLCipher$EVP_CIPHER$AES_128$CBC$PKCS5Padding");
+                "AES_128/CBC/PKCS5Padding", "OpenSSLEvpCipherAES$AES_128$CBC$PKCS5Padding");
         put("Alg.Alias.Cipher.AES_128/CBC/PKCS7Padding", "AES_128/CBC/PKCS5Padding");
 
         put("Alg.Alias.Cipher.PBEWithHmacSHA1AndAES_128", "AES_128/CBC/PKCS5PADDING");
@@ -385,14 +385,14 @@ public final class OpenSSLProvider extends Provider {
         put("Alg.Alias.Cipher.PBEWithHmacSHA512AndAES_128", "AES_128/CBC/PKCS5PADDING");
 
         putSymmetricCipherImplClass(
-                "AES_256/ECB/NoPadding", "OpenSSLCipher$EVP_CIPHER$AES_256$ECB$NoPadding");
+                "AES_256/ECB/NoPadding", "OpenSSLEvpCipherAES$AES_256$ECB$NoPadding");
         putSymmetricCipherImplClass(
-                "AES_256/ECB/PKCS5Padding", "OpenSSLCipher$EVP_CIPHER$AES_256$ECB$PKCS5Padding");
+                "AES_256/ECB/PKCS5Padding", "OpenSSLEvpCipherAES$AES_256$ECB$PKCS5Padding");
         put("Alg.Alias.Cipher.AES_256/ECB/PKCS7Padding", "AES_256/ECB/PKCS5Padding");
         putSymmetricCipherImplClass(
-                "AES_256/CBC/NoPadding", "OpenSSLCipher$EVP_CIPHER$AES_256$CBC$NoPadding");
+                "AES_256/CBC/NoPadding", "OpenSSLEvpCipherAES$AES_256$CBC$NoPadding");
         putSymmetricCipherImplClass(
-                "AES_256/CBC/PKCS5Padding", "OpenSSLCipher$EVP_CIPHER$AES_256$CBC$PKCS5Padding");
+                "AES_256/CBC/PKCS5Padding", "OpenSSLEvpCipherAES$AES_256$CBC$PKCS5Padding");
         put("Alg.Alias.Cipher.AES_256/CBC/PKCS7Padding", "AES_256/CBC/PKCS5Padding");
 
         put("Alg.Alias.Cipher.PBEWithHmacSHA1AndAES_256", "AES_256/CBC/PKCS5PADDING");
@@ -402,31 +402,31 @@ public final class OpenSSLProvider extends Provider {
         put("Alg.Alias.Cipher.PBEWithHmacSHA512AndAES_256", "AES_256/CBC/PKCS5PADDING");
 
         putSymmetricCipherImplClass("DESEDE/CBC/NoPadding",
-                "OpenSSLCipher$EVP_CIPHER$DESEDE$CBC$NoPadding");
+                "OpenSSLEvpCipherDESEDE$CBC$NoPadding");
         putSymmetricCipherImplClass("DESEDE/CBC/PKCS5Padding",
-                "OpenSSLCipher$EVP_CIPHER$DESEDE$CBC$PKCS5Padding");
+                "OpenSSLEvpCipherDESEDE$CBC$PKCS5Padding");
         put("Alg.Alias.Cipher.DESEDE/CBC/PKCS7Padding", "DESEDE/CBC/PKCS5Padding");
 
-        putSymmetricCipherImplClass("ARC4", "OpenSSLCipher$EVP_CIPHER$ARC4");
+        putSymmetricCipherImplClass("ARC4", "OpenSSLEvpCipherARC4");
         put("Alg.Alias.Cipher.ARCFOUR", "ARC4");
         put("Alg.Alias.Cipher.RC4", "ARC4");
         put("Alg.Alias.Cipher.1.2.840.113549.3.4", "ARC4");
         put("Alg.Alias.Cipher.OID.1.2.840.113549.3.4", "ARC4");
 
-        putSymmetricCipherImplClass("AES/GCM/NoPadding", "OpenSSLCipher$EVP_AEAD$AES$GCM");
+        putSymmetricCipherImplClass("AES/GCM/NoPadding", "OpenSSLAeadCipherAES$GCM");
         put("Alg.Alias.Cipher.GCM", "AES/GCM/NoPadding");
         put("Alg.Alias.Cipher.2.16.840.1.101.3.4.1.6", "AES/GCM/NoPadding");
         put("Alg.Alias.Cipher.2.16.840.1.101.3.4.1.26", "AES/GCM/NoPadding");
         put("Alg.Alias.Cipher.2.16.840.1.101.3.4.1.46", "AES/GCM/NoPadding");
         putSymmetricCipherImplClass(
-                "AES_128/GCM/NoPadding", "OpenSSLCipher$EVP_AEAD$AES$GCM$AES_128");
+                "AES_128/GCM/NoPadding", "OpenSSLAeadCipherAES$GCM$AES_128");
         putSymmetricCipherImplClass(
-                "AES_256/GCM/NoPadding", "OpenSSLCipher$EVP_AEAD$AES$GCM$AES_256");
+                "AES_256/GCM/NoPadding", "OpenSSLAeadCipherAES$GCM$AES_256");
 
         putSymmetricCipherImplClass("ChaCha20",
                 "OpenSSLCipherChaCha20");
         putSymmetricCipherImplClass("ChaCha20/Poly1305/NoPadding",
-                "OpenSSLCipher$EVP_AEAD$ChaCha20");
+                "OpenSSLAeadCipherChaCha20");
         put("Alg.Alias.Cipher.ChaCha20-Poly1305", "ChaCha20/Poly1305/NoPadding");
 
         /* === Mac === */
